pyproject.toml
[project]
name = "mental-models-simulator"
version = "0.2.0"
description = "Ground-truth qualitative simulator for everyday objects (benchmark-focused)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.4.0",
    "PyYAML>=6.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pytest>=8.4.1",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=3.5.0",
    "ruff>=0.6.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
sim = "simulator.cli.app:app"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = [
    "E402",  # Module level import not at top of file (common with docstrings)
    "E731",  # Do not assign a lambda expression
]

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true
strict_optional = true

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
